Split dashboards into multiple files.
