+++ templates/pages/building/building_core.html
@@ -25,7 +25,7 @@
       aria-selected="false"
     >Material</button>
   </li>
-  <li class="nav-item" role="presentation">
+  <!-- <li class="nav-item" role="presentation">
     <button
       class="nav-link"
       id="benchmarking-tab"
@@ -48,7 +48,7 @@
       aria-controls="savings"
       aria-selected="false"
     >Savings Potential</button>
-  </li>
+  </li> -->
 </ul>
 
 <div class="tab-content" id="myTabContent">
@@ -64,6 +64,31 @@
       id="assembly_dashboard"
       hx-get="{% url 'dashboard' %}?model=building&id={{building_id}}&dashboard_type=assembly&simulation={{simulation}}"
       hx-trigger="load, refresh"
+      hx-swap="innerHTML"
+    >
+      <div class="text-center">
+        <img
+          alt="Result loading..."
+          class="htmx-indicator"
+          width="150"
+          src="https://htmx.org/img/bars.svg"
+        />
+      </div>
+    </div>
+  </div>
+
+
+  <!-- MATERIAL -->
+  <div
+    class="tab-pane fade p-3"
+    id="material"
+    role="tabpanel"
+    aria-labelledby="material-tab"
+  >
+    <div
+      id="material_dashboard"
+      hx-get="{% url 'dashboard' %}?model=building&id={{building_id}}&dashboard_type=material&simulation={{simulation}}"
+      hx-trigger="shown.bs.tab from:#material-tab, refresh"
       hx-swap="innerHTML"
     >
       <div class="text-center">
